+++ playground/nuxt.config.ts
@@ -2,7 +2,7 @@ import { defineNuxtConfig } from 'nuxt/config'
 
 export default defineNuxtConfig({
 	modules: ['nuxt-sharekit'],
-	devtools: { enabled: true },
+	devtools: { enabled: false },
 	compatibilityDate: '2026-07-21',
 	shareKit: {
 		componentPrefix: 'Share',